+++ NEWS.md
@@ -1,4 +1,5 @@
 # parquetize 0.1.0
 
 * Added `csv_to_parquet()` function to convert csv files to parquet format.
+* Added `parquetize_example()` function to get path to package data examples.
 * Added a `NEWS.md` file to track changes to the package.